Remove old code in gammapy.image
