+++ gammapy/image/models/shapes.py
